threat-api-tracker/test_target/requirements.txt
Flask==2.0.1
PyYAML==5.1
simplejson==3.17.2
urllib3==1.26.5
itsdangerous==2.0.1
cryptography==3.4.7
MarkupSafe==2.0.1
python-multipart==0.0.5
